+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/manipulators/PlaybackEventDataManipulator.kt
@@ -16,6 +16,8 @@ internal class PlaybackEventDataManipulator(
             // IVS player only supports HLS, thus we hardcode it here
             data.streamFormat = StreamFormat.HLS.value
 
+            data.m3u8Url = extractM3u8Url(player)
+
             // It cannot be distinguished between a live stream and a stream not being loaded
             // on IVS, thus we don't use the isLive property from the SourceMetadata as fallback
             data.isLive = Utils.isPlaybackLive(player.duration)
@@ -29,5 +31,19 @@ internal class PlaybackEventDataManipulator(
         } catch (e: Exception) {
             Log.e("PlaybackDataManipulator", "Something went wrong while setting playback event data, e: ${e.message}", e)
         }
+    }
+
+    // IVS Player only supports path api starting with version 1.23
+    private fun extractM3u8Url(player: Player): String? {
+        runCatching {
+            val splitted = player.version.split(".")
+            val majorVersion = splitted[0].toInt()
+            val minorVersion = splitted[1].toInt()
+
+            if (majorVersion >= 1 && minorVersion >= 23) {
+                return player.path
+            }
+            return null
+        }.getOrNull() ?: return null
     }
 }