+++ collector-amazon-ivs/src/test/java/android/util/Log.java
@@ -22,4 +22,9 @@ public class Log {
         System.out.println("ERROR: " + tag + ": " + msg);
         return 0;
     }
+
+    public static int e(String tag, String msg, Throwable t) {
+        System.out.println("ERROR: " + tag + ": " + msg + " " + t.toString());
+        return 0;
+    }
 }